dom/indexedDB/test/test_unique_index_update.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
    function testSteps()
    {
      let request = mozIndexedDB.open(window.location.pathname);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield;

      let db = event.target.result;

      request = db.setVersion("1");
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      for each (let name in ["1", "2"]) {
        let objectStore =
          db.createObjectStore(name, { keyPath: "id",
                                       autoIncrement: name == "2" });
        objectStore.createIndex("1", "index", { unique: true });

        for (let i = 0; i < 10; i++) {
          objectStore.add({ id: i, index: i });
        }
      }

      event.target.transaction.oncomplete = continueToNextStep;
      yield;

      for each (let name in ["1", "2"]) {
        objectStore = db.transaction(name, IDBTransaction.READ_WRITE)
                        .objectStore(name);

        request = objectStore.put({ id: 5, index: 6 });
        request.onsuccess = unexpectedSuccessHandler;
        request.onerror = new ExpectError(IDBDatabaseException.CONSTRAINT_ERR);
        event = yield;

        event.preventDefault();

        let keyRange = IDBKeyRange.only(5);

        objectStore.index("1").openCursor(keyRange).onsuccess = function(event) {
          let cursor = event.target.result;
          ok(cursor, "Must have a cursor here");

          is(cursor.value.index, 5, "Still have the right index value");

          cursor.value.index = 6;

          request = cursor.update(cursor.value);
          request.onsuccess = unexpectedSuccessHandler;
          request.onerror =
            new ExpectError(IDBDatabaseException.CONSTRAINT_ERR);
        };

        yield;
      }

      finishTest();
      yield;
    }

  </script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
